Rename SupportedCommand to CommandType
