+++ src/org/traccar/web/WebServer.java
@@ -43,7 +43,7 @@ import org.traccar.api.resource.UserResource;
 import org.traccar.api.resource.GroupResource;
 import org.traccar.api.resource.DeviceResource;
 import org.traccar.api.resource.PositionResource;
-import org.traccar.api.resource.SupportedCommandResource;
+import org.traccar.api.resource.CommandTypeResource;
 import org.traccar.helper.Log;
 
 import javax.naming.InitialContext;
@@ -148,7 +148,7 @@ public class WebServer {
         resourceConfig.register(CorsResponseFilter.class);
         resourceConfig.registerClasses(ServerResource.class, SessionResource.class, CommandResource.class,
                 GroupPermissionResource.class, DevicePermissionResource.class, UserResource.class,
-                GroupResource.class, DeviceResource.class, PositionResource.class, SupportedCommandResource.class);
+                GroupResource.class, DeviceResource.class, PositionResource.class, CommandTypeResource.class);
         servletHandler.addServlet(new ServletHolder(new ServletContainer(resourceConfig)), "/*");
 
         handlers.addHandler(servletHandler);
